+++ src/main/java/org/folio/entlinks/service/consortium/propagation/ConsortiumAuthoritySourceFilePropagationService.java
@@ -6,7 +6,7 @@ import org.folio.entlinks.service.authority.AuthoritySourceFileService;
 import org.folio.entlinks.service.consortium.ConsortiumTenantsService;
 import org.folio.entlinks.service.consortium.propagation.model.AuthoritySourceFilePropagationData;
 import org.folio.spring.FolioExecutionContext;
-import org.folio.spring.service.SystemUserScopedExecutionService;
+import org.folio.spring.scope.FolioExecutionContextService;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
@@ -20,7 +20,7 @@ public class ConsortiumAuthoritySourceFilePropagationService
   public ConsortiumAuthoritySourceFilePropagationService(
       @Qualifier("propagationAuthoritySourceFileService") AuthoritySourceFileService sourceFileService,
       ConsortiumTenantsService tenantsService,
-      SystemUserScopedExecutionService executionService,
+      FolioExecutionContextService executionService,
       FolioExecutionContext folioExecutionContext) {
     super(tenantsService, executionService, folioExecutionContext);
     this.sourceFileService = sourceFileService;