+++ src/test/java/org/folio/entlinks/integration/kafka/AuthorityEventListenerTest.java
@@ -1,6 +1,8 @@
 package org.folio.entlinks.integration.kafka;
 
 import static java.util.Collections.singletonList;
+import static org.folio.spring.integration.XOkapiHeaders.URL;
+import static org.folio.spring.integration.XOkapiHeaders.USER_ID;
 import static org.folio.support.MockingTestUtils.mockBatchFailedHandling;
 import static org.folio.support.MockingTestUtils.mockBatchSuccessHandling;
 import static org.mockito.ArgumentMatchers.any;
@@ -10,6 +12,7 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
@@ -20,8 +23,7 @@ import org.folio.entlinks.domain.dto.AuthorityDto;
 import org.folio.entlinks.domain.dto.Metadata;
 import org.folio.entlinks.integration.dto.event.AuthorityDomainEvent;
 import org.folio.entlinks.service.messaging.authority.InstanceAuthorityLinkUpdateService;
-import org.folio.spring.integration.XOkapiHeaders;
-import org.folio.spring.service.SystemUserScopedExecutionService;
+import org.folio.spring.scope.FolioExecutionContextService;
 import org.folio.spring.testing.type.UnitTest;
 import org.folio.spring.tools.batch.MessageBatchProcessor;
 import org.folio.support.TestDataUtils;
@@ -39,7 +41,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 class AuthorityEventListenerTest {
 
   @Mock
-  private SystemUserScopedExecutionService executionService;
+  private FolioExecutionContextService executionService;
   @Mock
   private InstanceAuthorityLinkUpdateService instanceAuthorityLinkUpdateService;
   @Mock
@@ -53,7 +55,7 @@ class AuthorityEventListenerTest {
 
   @BeforeEach
   void setUp() {
-    when(executionService.executeSystemUserScoped(any(), anyMap(), any())).thenAnswer(invocation -> {
+    when(executionService.execute(any(), anyMap(), any(Callable.class))).thenAnswer(invocation -> {
       var argument = invocation.getArgument(2, Callable.class);
       return argument.call();
     });
@@ -67,7 +69,8 @@ class AuthorityEventListenerTest {
     var oldRecord = new AuthorityDto().id(authId);
     var userId = UUID.randomUUID().toString();
     var headers = new RecordHeaders();
-    headers.add(XOkapiHeaders.USER_ID, userId.getBytes());
+    headers.add(USER_ID, userId.getBytes());
+    headers.add(URL, "http://localhost:8081".getBytes());
     var event = TestDataUtils.authorityEvent(type, newRecord, oldRecord);
 
     mockBatchSuccessHandling(messageBatchProcessor);
@@ -77,7 +80,9 @@ class AuthorityEventListenerTest {
 
     listener.handleEvents(singletonList(consumerRecord));
 
-    verify(executionService).executeSystemUserScoped(any(), eq(Map.of(XOkapiHeaders.USER_ID, List.of(userId))), any());
+    Map<String, Collection<String>> expectedHeaders =
+      Map.of(USER_ID, List.of(userId), URL, List.of("http://localhost:8081"));
+    verify(executionService).execute(any(), eq(expectedHeaders), any(Callable.class));
     verify(instanceAuthorityLinkUpdateService).handleAuthoritiesChanges(singletonList(event));
   }
 
@@ -89,11 +94,14 @@ class AuthorityEventListenerTest {
     var meta = new Metadata().updatedByUserId(updatedByUserId);
     var newRecord = new AuthorityDto().id(authId).metadata(meta);
     var oldRecord = new AuthorityDto().id(authId).metadata(meta.updatedByUserId(updatedByUserId));
+    var headers = new RecordHeaders();
+    headers.add(URL, "http://localhost:8081".getBytes());
     var event = TestDataUtils.authorityEvent(type, newRecord, oldRecord);
 
     mockBatchSuccessHandling(messageBatchProcessor);
     when(consumerRecord.key()).thenReturn(authId.toString());
     when(consumerRecord.value()).thenReturn(event);
+    when(consumerRecord.headers()).thenReturn(headers);
 
     listener.handleEvents(singletonList(consumerRecord));
 
@@ -105,11 +113,14 @@ class AuthorityEventListenerTest {
     var authId = UUID.randomUUID();
     var newRecord = new AuthorityDto().id(authId);
     var oldRecord = new AuthorityDto().id(authId);
+    var headers = new RecordHeaders();
+    headers.add(URL, "http://localhost:8081".getBytes());
     var event = TestDataUtils.authorityEvent("UPDATE", newRecord, oldRecord);
 
     mockBatchFailedHandling(messageBatchProcessor, new RuntimeException("test message"));
     when(consumerRecord.key()).thenReturn(authId.toString());
     when(consumerRecord.value()).thenReturn(event);
+    when(consumerRecord.headers()).thenReturn(headers);
 
     listener.handleEvents(singletonList(consumerRecord));