+++ NEWS.md
@@ -14,7 +14,7 @@
 * Add validation for Authority record creation during Data Import ([MODELINKS-417](https://folio-org.atlassian.net/browse/MODELINKS-417))
 
 ### Tech Dept
-* Description ([ISSUE](https://folio-org.atlassian.net/browse/ISSUE))
+* Remove use of deprecated folio-spring-system-user ([MODELINKS-339](https://folio-org.atlassian.net/browse/MODELINKS-339))
 
 ### Dependencies
 * Bump `LIB_NAME` from `OLD_VERSION` to `NEW_VERSION`